rf: move queue populator out of extensions/
As a first step to make it generic, put the queue populator code in
the lib/queuePopulator directory, and the executable file in
bin/queuePopulator.js.
