+++ lib/queuePopulator/QueuePopulator.js
@@ -2,10 +2,9 @@ const async = require('async');
 
 const Logger = require('werelogs').Logger;
 
-const zookeeper = require('../../../lib/clients/zookeeper');
-const BackbeatProducer = require('../../../lib/BackbeatProducer');
-const ProvisionDispatcher =
-          require('../../../lib/provisioning/ProvisionDispatcher');
+const zookeeper = require('../clients/zookeeper');
+const BackbeatProducer = require('../BackbeatProducer');
+const ProvisionDispatcher = require('../provisioning/ProvisionDispatcher');
 const RaftLogReader = require('./RaftLogReader');
 const BucketFileLogReader = require('./BucketFileLogReader');
