+++ package.json
@@ -4,7 +4,7 @@
   "description": "Asynchronous queue and job manager",
   "main": "index.js",
   "scripts": {
-    "queue_populator": "node extensions/replication/queuePopulator/task.js",
+    "queue_populator": "node bin/queuePopulator.js",
     "queue_processor": "node extensions/replication/queueProcessor/task.js",
     "test": "mocha --recursive tests/unit",
     "ft_test": "mocha --recursive tests/functional",
